requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.36.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
idna==2.6
Markdown==3.0.1
oauthlib==2.1.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python3-openid==3.1.0
pytz==2018.7
requests==2.18.4
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.22
